requirements.txt
boto3==1.24.70
botocore==1.27.70
Django==4.1.1
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-froala-editor==4.0.14
django-storages==1.13.1
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.21.0
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
urllib3==1.26.12
virtualenv==20.16.5
virtualenvwrapper-win==1.2.7
Werkzeug==2.2.2
whitenoise==6.2.0